requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
Wand==0.6.7
whitenoise==5.3.0